fix: ModInverse extension in BigIntegerExtensions
